requirements.txt
chardet==5.2.0
pandas==2.2.2
peewee==3.17.5
Pillow==10.3.0
pip==24.0
pyodbc==5.1.0
python-dotenv==1.0.1
tkcalendar==1.6.1
